requirements.txt
streamlit
google-cloud-storage
google-generativeai
python-dotenv
